rename BenefitTile -> BenefitTiles
